Kaggle Titanic Dataset-Machine Learning from Disaster
We got access to two similar datasets  through Kaggle that included passenger information like name, age, gender, socio-economic class, etc. One dataset is titled `train.csv` and the other is titled `test.csv`.

Train.csv contains the details of a subset of the passengers on board (891 to be exact) and importantly, will reveal whether they survived or not, also known as the “ground truth”.

The `test.csv` dataset contains similar information but does not disclose the “ground truth” for each passenger. It’s your job to predict these outcomes.

Using the patterns we find in the train.csv data, predict whether the other 418 passengers on board (found in test.csv) survived.